+++ src/lob_asset_management/controller/metric.clj
@@ -7,9 +7,23 @@
   [endpoint]
   (let [{:metric/keys [api-call] :as metric} (io.f-in/get-file-by-entity :metric)
         api-call' (->> {:endpoint endpoint
-                        :when     (aux.t/date-time-str)}
+                        :when     (aux.t/current-datetime->str)}
                        list
                        (concat (or api-call []))
                        (assoc metric :metric/api-call))
         ]
     (io.f-out/metric-file api-call')))
+
+(defn total-api-calls
+  ([]
+   (total-api-calls (io.f-in/get-file-by-entity :metric)))
+  ([{:metric/keys [api-call]}]
+   (let [today-keyword (aux.t/current-date->keyword)
+         calls-today (->> api-call
+                          (map #(assoc % :when-date (aux.t/clj-date->date-keyword (:when %))))
+                          (group-by :when-date)
+                          today-keyword
+                          count)]
+     {:today       calls-today
+      :total       (count api-call)
+      :daily-limit 500})))